async api works

+++ src/main/java/com/spoloborota/calculator/service/ICalculationServiceAsync.java
@@ -5,13 +5,14 @@ import com.spoloborota.calculator.exception.WrongExpressionException;
 import com.spoloborota.calculator.exception.WrongOperationException;
 import org.springframework.data.domain.Page;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.concurrent.CompletableFuture;
 
 public interface ICalculationServiceAsync {
     CompletableFuture<Double> calculateExpression(String expr) throws WrongExpressionException;
-    CompletableFuture<Long> countByDate(LocalDate date);
-    CompletableFuture<Long> countContainsOp(String operation) throws WrongOperationException;
+    CompletableFuture<BigInteger> countByDate(LocalDate date);
+    CompletableFuture<BigInteger> countContainsOp(String operation) throws WrongOperationException;
     CompletableFuture<Page<Calculation>> listByDate(LocalDate date, int page, int size);
     CompletableFuture<Page<Calculation>> listContainsOp(String operation, int page, int size) throws WrongOperationException;
     CompletableFuture<Double> popularNumber();
